Change withMotionDetail to useMotionDetail
