+++ src/screen/DashboardScreen.tsx
@@ -6,6 +6,7 @@ import {useAccounts} from 'context/AccountsContext';
 import {ChainApiContext} from 'context/ChainApiContext';
 import {NetworkContext} from 'context/NetworkContext';
 import {CouncilSummaryTeaser} from 'layout/CouncilSummaryTeaser';
+import {MotionTeaser} from 'layout/MotionTeaser';
 import {ReferendaSummaryTeaser} from 'layout/ReferendaSummaryTeaser';
 import TipsSummaryTeaser from 'layout/tips/TipsSummaryTeaser';
 import {TreasurySummaryTeaser} from 'layout/TreasurySummaryTeaser';
